write all Q3 parts as funcitons
I seperated out everything we need to do just once and everything we
need to do in the iterations. Anything in for the iterations is now
written as a function. I also made it so that none of the funcitons
reference global objects. Now everythign the function uses is fed into
the function as a parameter.